Added more pep8 related fixes
Fixed rebase issue in cps test
