+++ lib/ansible/module_utils/hpe3par.py
@@ -1,0 +1,84 @@
+storage_system_spec = {
+    "storage_system_ip": {
+        "required": True,
+        "type": "str"
+    },
+    "storage_system_username": {
+        "required": True,
+        "type": "str",
+        "no_log": True
+    },
+    "storage_system_password": {
+        "required": True,
+        "type": "str",
+        "no_log": True
+    },
+    "secure": {
+        "type": "bool",
+        "default": False
+    }
+}
+
+
+def cpg_argument_spec():
+    spec = {
+        "state": {
+            "required": True,
+            "choices": ['present', 'absent'],
+            "type": 'str'
+        },
+        "cpg_name": {
+            "required": True,
+            "type": "str"
+        },
+        "domain": {
+            "type": "str"
+        },
+        "growth_increment": {
+            "type": "float",
+            "default": -1.0
+        },
+        "growth_increment_unit": {
+            "type": "str",
+            "choices": ['TiB', 'GiB', 'MiB'],
+            "default": 'GiB'
+        },
+        "growth_limit": {
+            "type": "float",
+            "default": -1.0
+        },
+        "growth_limit_unit": {
+            "type": "str",
+            "choices": ['TiB', 'GiB', 'MiB'],
+            "default": 'GiB'
+        },
+        "growth_warning": {
+            "type": "float",
+            "default": -1.0
+        },
+        "growth_warning_unit": {
+            "type": "str",
+            "choices": ['TiB', 'GiB', 'MiB'],
+            "default": 'GiB'
+        },
+        "raid_type": {
+            "required": False,
+            "type": "str",
+            "choices": ['R0', 'R1', 'R5', 'R6']
+        },
+        "set_size": {
+            "required": False,
+            "type": "int",
+            "default": -1
+        },
+        "high_availability": {
+            "type": "str",
+            "choices": ['PORT', 'CAGE', 'MAG']
+        },
+        "disk_type": {
+            "type": "str",
+            "choices": ['FC', 'NL', 'SSD']
+        }
+    }
+    spec.update(storage_system_spec)
+    return spec
