+++ lib/ansible/utils/module_docs_fragments/hpe3par.py
@@ -23,20 +23,20 @@ class ModuleDocFragment(object):
 options:
     storage_system_ip:
       description:
-        - "The storage system IP address."
+        - The storage system IP address.
       required: true
     storage_system_password:
       description:
-        - "The storage system password."
+        - The storage system password.
       required: true
     storage_system_username:
       description:
-        - "The storage system user name."
+        - The storage system user name.
       required: true
 
 requirements:
-  - "3PAR OS - 3.2.2 MU6, 3.3.1 MU1"
-  - "Ansible - 2.4"
-  - "hpe3par_sdk 1.0.0"
-  - "WSAPI service should be enabled on the 3PAR storage array."
+  - 3PAR OS - 3.2.2 MU6, 3.3.1 MU1
+  - Ansible - 2.7
+  - hpe3par_sdk
+  - WSAPI service should be enabled on the 3PAR storage array.
     '''
